+++ src/SendCallback/CallbackInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Facile\SentryModule\SendCallback;
+
+interface CallbackInterface
+{
+    /**
+     * @param array $data
+     * @return array
+     */
+    public function __invoke(array $data);
+}