+++ src/Transport/TransportInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Facile\SentryModule\Transport;
+
+interface TransportInterface
+{
+    /**
+     * @param \Raven_Client $client
+     * @param array $data
+     */
+    public function __invoke(\Raven_Client $client, array $data);
+}